+++ internal/update/reexec_unix.go
@@ -1,0 +1,9 @@
+//go:build !windows
+
+package update
+
+import "syscall"
+
+func reexecSelf(exe string, args, env []string) error {
+	return syscall.Exec(exe, args, env)
+}